error when trying to insert strings w/ special characters into db
